+++ src/main/java/module-info.java
@@ -1,22 +1,3 @@
-import no.ssb.rawdata.api.state.StatePersistenceInitializer;
-import no.ssb.rawdata.api.storage.RawdataClientInitializer;
-
 module no.ssb.rawdata.api {
-    requires no.ssb.config;
-    requires no.ssb.service.provider.api;
-    requires io.reactivex.rxjava2;
-    requires org.reactivestreams;
-    requires jackson.annotations;
-    requires com.fasterxml.jackson.core;
-    requires com.fasterxml.jackson.databind;
-    requires java.sql;
-    requires org.slf4j;
-
-    uses StatePersistenceInitializer;
-    uses RawdataClientInitializer;
-
-    exports no.ssb.rawdata.api.persistence;
-    exports no.ssb.rawdata.api.state;
-    exports no.ssb.rawdata.api.storage;
-    exports no.ssb.rawdata.api.util;
+    exports no.ssb.rawdata.api;
 }
