+++ src/main/java/no/ssb/rawdata/api/RawdataClient.java
@@ -1,0 +1,9 @@
+package no.ssb.rawdata.api;
+
+public interface RawdataClient {
+
+    RawdataProducer producer(String topic);
+
+    RawdataConsumer consumer(String topic, String subscription);
+
+}
