+++ src/main/java/no/ssb/rawdata/api/RawdataMessage.java
@@ -1,0 +1,9 @@
+package no.ssb.rawdata.api;
+
+public interface RawdataMessage {
+
+    RawdataMessageId id();
+
+    RawdataMessageContent content();
+
+}
