+++ src/main/java/no/ssb/rawdata/api/RawdataMessageContent.java
@@ -1,0 +1,12 @@
+package no.ssb.rawdata.api;
+
+import java.util.Set;
+
+public interface RawdataMessageContent {
+
+    String externalId();
+
+    Set<String> keys();
+
+    byte[] get(String key);
+}
