+++ src/main/java/no/ssb/rawdata/api/RawdataMessageId.java
@@ -1,0 +1,4 @@
+package no.ssb.rawdata.api;
+
+public interface RawdataMessageId {
+}
