+++ src/main/java/no/ssb/rawdata/api/RawdataProducer.java
@@ -1,0 +1,19 @@
+package no.ssb.rawdata.api;
+
+import java.util.List;
+
+public interface RawdataProducer {
+
+    /**
+     * Get the last externalId in the stream
+     *
+     * @return
+     */
+    String lastExternalId();
+
+    void buffer(RawdataMessageContent content);
+
+    List<? extends RawdataMessageId> publish(List<String> externalIds);
+
+    List<? extends RawdataMessageId> publish(String... externalIds);
+}
