--- src/main/java/no/ssb/rawdata/api/persistence/Disposable.java
@@ -1,13 +1,0 @@
-package no.ssb.rawdata.api.persistence;
-
-import java.io.Closeable;
-
-public interface Disposable extends Closeable {
-
-    void cancel();
-
-    void dispose();
-
-    @Override
-    void close();
-}
