--- src/main/java/no/ssb/rawdata/api/persistence/PersistenceException.java
@@ -1,24 +1,0 @@
-package no.ssb.rawdata.api.persistence;
-
-public class PersistenceException extends RuntimeException {
-
-    public PersistenceException() {
-        super();
-    }
-
-    public PersistenceException(String message) {
-        super(message);
-    }
-
-    public PersistenceException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public PersistenceException(Throwable cause) {
-        super(cause);
-    }
-
-    protected PersistenceException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
