--- src/main/java/no/ssb/rawdata/api/persistence/Transaction.java
@@ -1,31 +1,0 @@
-package no.ssb.rawdata.api.persistence;
-
-import java.util.concurrent.CompletableFuture;
-
-public interface Transaction extends AutoCloseable {
-
-    CompletableFuture<TransactionStatistics> commit();
-
-    CompletableFuture<TransactionStatistics> cancel();
-
-    @Override
-    default void close() {
-        boolean committed = false;
-        try {
-            commit().join();
-            committed = true;
-        } catch (Throwable t) {
-            if (t instanceof RuntimeException) {
-                throw (RuntimeException) t;
-            }
-            if (t instanceof Error) {
-                throw (Error) t;
-            }
-            throw new RuntimeException(t);
-        } finally {
-            if (!committed) {
-                cancel().join();
-            }
-        }
-    }
-}
