--- src/main/java/no/ssb/rawdata/api/state/StatePersistence.java
@@ -1,22 +1,0 @@
-package no.ssb.rawdata.api.state;
-
-import io.reactivex.Flowable;
-import io.reactivex.Maybe;
-import io.reactivex.Single;
-
-import java.io.Closeable;
-import java.util.List;
-
-public interface StatePersistence extends Closeable {
-
-    Single<Boolean> trackCompletedPositions(String namespace, List<String> completedPositions);
-
-    Maybe<String> getFirstPosition(String namespace);
-
-    Maybe<String> getLastPosition(String namespace);
-
-    Maybe<String> getOffsetPosition(String namespace, String fromPositon, int offset);
-
-    Flowable<CompletedPosition> readPositions(String namespace, String fromPosition, String toPosition);
-
-}
