--- src/main/java/no/ssb/rawdata/api/state/StatePersistenceInitializer.java
@@ -1,7 +1,0 @@
-package no.ssb.rawdata.api.state;
-
-import no.ssb.service.provider.api.ProviderInitializer;
-
-public interface StatePersistenceInitializer extends ProviderInitializer<StatePersistence> {
-
-}
