--- src/main/java/no/ssb/rawdata/api/storage/RawdataAPI.java
@@ -1,9 +1,0 @@
-package no.ssb.rawdata.api.storage;
-
-public interface RawdataAPI {
-
-    RawdataProducer producer(String stream);
-
-    RawdataConsumer consumer(String stream, String subscription, SubscriptionMode mode);
-
-}
