--- src/main/java/no/ssb/rawdata/api/storage/RawdataClientInitializer.java
@@ -1,7 +1,0 @@
-package no.ssb.rawdata.api.storage;
-
-import no.ssb.service.provider.api.ProviderInitializer;
-
-public interface RawdataClientInitializer extends ProviderInitializer<RawdataClient> {
-
-}
