--- src/main/java/no/ssb/rawdata/api/storage/RawdataMessage.java
@@ -1,9 +1,0 @@
-package no.ssb.rawdata.api.storage;
-
-public interface RawdataMessage {
-
-    RawdataMessageId id();
-
-    RawdataMessageContent read();
-
-}
