--- src/main/java/no/ssb/rawdata/api/storage/RawdataMessageContent.java
@@ -1,9 +1,0 @@
-package no.ssb.rawdata.api.storage;
-
-public interface RawdataMessageContent {
-
-    String position();
-
-    byte[] read(String filename);
-
-}
