--- src/main/java/no/ssb/rawdata/api/storage/RawdataMessageId.java
@@ -1,5 +1,0 @@
-package no.ssb.rawdata.api.storage;
-
-public interface RawdataMessageId {
-
-}
